refactor: home page, move code from home fetch posts to home page
