--- src/components/postsPage/PostItem.tsx
@@ -1,88 +1,0 @@
-import type { FC, ReactNode } from "react"
-import type { PostMenuItemsType, PostWithAuthor } from "./types"
-import Link from "next/link"
-import dayjs from "dayjs"
-import relativeTime from "dayjs/plugin/relativeTime"
-import { PostOptionMenu } from "./PostOptionMenu"
-import { Icon } from "../Icon"
-import { PostQuoteItem } from "./PostQuoteItem"
-import { PostPoll } from "./PostPoll"
-import { useTimeLeft } from "~/hooks/useTimeLeft"
-import { PostTitle } from "./PostTitle"
-import { ProfileAvatarImageUrl } from "../profile/ProfileAvatarImageUrl"
-
-dayjs.extend(relativeTime)
-
-export type ClickCapture = {
-	action: "deletePost" | "navigation" | "vote"
-	choiceId?: number
-}
-
-export const PostItem: FC<{
-	postWithUser: PostWithAuthor
-	menuItemsType: PostMenuItemsType
-	onClickCapture: (clickCapture: ClickCapture) => void
-	footer?: ReactNode
-}> = ({ postWithUser, menuItemsType, onClickCapture, footer }) => {
-	const useTime = useTimeLeft(postWithUser.post.createdAt, postWithUser.post.poll?.endDate)
-	const { post, author } = postWithUser
-	return (
-		<li
-			className="cursor-pointer rounded-lg py-2 hover:bg-gray-100"
-			onClick={() => {
-				onClickCapture({
-					action: "navigation",
-				})
-			}}
-		>
-			<main className="flex">
-				<ProfileAvatarImageUrl src={author.profileImageUrl} />
-				<div className="w-full pl-2">
-					<PostTitle author={author} createdAt={post.createdAt} />
-					<span>{post.content}</span>
-					{post.poll && (
-						<PostPoll
-							pollTimeLeft={useTime}
-							poll={post.poll}
-							pollEndTime={post.poll.endDate}
-							onClickVote={(id) =>
-								onClickCapture({
-									action: "vote",
-									choiceId: id,
-								})
-							}
-						/>
-					)}
-					{post.quotedPost && (
-						<Link
-							onClick={(e) => e.stopPropagation()}
-							href={`/post/${post.quotedPost.author.username}/status/${post.quotedPost.post.id}`}
-						>
-							<PostQuoteItem
-								author={author}
-								createdAt={post.createdAt}
-								message={post.content}
-							/>
-						</Link>
-					)}
-					{footer}
-				</div>
-				{menuItemsType !== "view" && (
-					<div className="group relative flex h-12 w-1/12 justify-center rounded-full hover:bg-gray-200">
-						<Icon iconKind="optionDots" />
-						<div className="hidden group-hover:block">
-							<PostOptionMenu
-								menuItemsType={menuItemsType}
-								onMenuItemClick={(_) =>
-									onClickCapture({
-										action: "deletePost",
-									})
-								}
-							/>
-						</div>
-					</div>
-				)}
-			</main>
-		</li>
-	)
-}
