+++ src/pages/home.tsx
@@ -5,25 +5,19 @@ import toast from "react-hot-toast"
 import { CONFIG } from "~/config"
 import { clerkClient, getAuth } from "@clerk/nextjs/server"
 import { type User } from "@clerk/nextjs/dist/api"
-import { Icon } from "~/components/Icon"
-import { PrimaryButton } from "~/components/styledHTMLElements/StyledButtons"
-import { type PostContent } from "~/components/homePage/types"
-import { useReducer, useState } from "react"
-import { CreatePoll } from "~/components/homePage/CreatePoll"
-import { pollLengthReducer } from "~/reducers/pollLengthReducer"
-import { pollChoicesReducer } from "~/reducers/pollChoicesReducer"
-import { ProfileAvatarImageUrl } from "~/components/profile/ProfileAvatarImageUrl"
+import { useRef, useState } from "react"
 import { type UserToFollow, WhoToFollow } from "~/features/whoToFollow"
 import { whoToFollow } from "~/server/features/whoToFollow"
-import { HomeFetchPosts } from "~/features/homePage/components/HomeFetchPosts"
-
-const INIT_POLL_LENGTH = {
-	days: 1,
-	hours: 0,
-	minutes: 0,
-}
-
-const INIT_POLL_CHOICES = ["", ""]
+import { type PostWithAuthor } from "~/components/post/types"
+import { type ClickCapture, PostItem } from "~/components/post/PostItem"
+import { LoadingPage } from "~/components/LoadingPage"
+import { useGetHomePosts } from "~/features/homePage/hooks"
+import { PostQuotePopUp } from "~/components/postsPage/PostQuotePopUp"
+import { PostContentSelector } from "~/components/post/PostContentSelector"
+import { PostFooter } from "~/components/postsPage/PostFooter"
+import { useRouter } from "next/router"
+import { ConnectorCreatePost } from "~/features/homePage"
+import { getPostMenuItemsType } from "~/hooks/getPostMenuItemsType"
 
 export const getServerSideProps: GetServerSideProps = async (props) => {
 	const { userId } = getAuth(props.req)
@@ -50,73 +44,16 @@ export const getServerSideProps: GetServerSideProps = async (props) => {
 }
 
 const Home: NextPage<{ user: User; usersToFollow: UserToFollow[] }> = ({ user, usersToFollow }) => {
-	//fetch asap
-	const posts = api.posts.getAllByAuthorId.useQuery(user.id)
-	const [postContent, setPostContent] = useState<PostContent>({
-		message: "",
-	})
+	const router = useRouter()
 
-	const [pollLengthState, pollLengthDispatch] = useReducer(pollLengthReducer, INIT_POLL_LENGTH)
-	const [pollChoicesState, pollChoicesDispatch] = useReducer(
-		pollChoicesReducer,
-		INIT_POLL_CHOICES
+	const [quotePopUp, setQuotePopUp] = useState<PostWithAuthor | null>(null)
+	const [quoteMessage, setQuoteMessage] = useState<string>()
+
+	const ulRef = useRef<HTMLUListElement>(null)
+
+	const { isLoading, posts, refetch } = useGetHomePosts(
+		ulRef.current && ulRef.current.scrollHeight - ulRef.current.offsetTop
 	)
-
-	const { mutate } = api.posts.createPost.useMutation({
-		onSuccess: async () => {
-			await posts.refetch()
-		},
-		onError: () => {
-			toast.error("Failed to create post! Please try again later", {
-				duration: CONFIG.TOAST_ERROR_DURATION_MS,
-			})
-		},
-	})
-
-	const handleCreatePost = () => {
-		if (postContent.poll) {
-			const notNullChoices = [...pollChoicesState].filter((choice) => {
-				if (choice) {
-					return choice
-				}
-			})
-			setPostContent({
-				...postContent,
-				poll: {
-					choices: notNullChoices,
-					length: pollLengthState,
-				},
-			})
-			mutate({
-				message: postContent.message,
-				poll: { choices: notNullChoices, length: pollLengthState },
-			})
-		} else {
-			mutate(postContent)
-		}
-	}
-
-	const handleRemovePoll = () => {
-		setPostContent({
-			...postContent,
-			poll: undefined,
-		})
-	}
-
-	const handlePollIconClick = () => {
-		if (postContent.poll) {
-			handleRemovePoll()
-		} else {
-			setPostContent({
-				...postContent,
-				message: postContent.message ?? "",
-				poll: {
-					choices: INIT_POLL_CHOICES,
-					length: pollLengthState,
-				},
-			})
-		}
-	}
 
 	const addUserToFollow = api.follow.addUserToFollow.useMutation({
 		onSuccess: (result) => {
@@ -128,6 +65,127 @@ const Home: NextPage<{ user: User; usersToFollow: UserToFollow[] }> = ({ user, u
 			})
 		},
 	})
+
+	const deletePost = api.posts.deletePost.useMutation({
+		onSuccess: async () => {
+			toast.success("Post Deleted!")
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to delete post! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	const quotePost = api.posts.createQuotedPost.useMutation({
+		onSuccess: async () => {
+			setQuotePopUp(null)
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to quote post! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	const likePost = api.profile.setPostLiked.useMutation({
+		onSuccess: async () => {
+			toast.success("Post Liked!")
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to like post! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	const unlikePost = api.profile.setPostUnliked.useMutation({
+		onSuccess: async () => {
+			toast.success("Post Unliked!")
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to unlike post! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	const forwardPost = api.posts.forwardPost.useMutation({
+		onSuccess: async () => {
+			toast.success("Post Forwarded!")
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to forward post! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	const removePostForward = api.posts.removePostForward.useMutation({
+		onSuccess: async () => {
+			toast.success("Delete Post Forward!")
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to delete forward! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	const pollVote = api.profile.votePostPoll.useMutation({
+		onSuccess: async () => {
+			toast.success("Voted!")
+			await refetch()
+		},
+		onError: () => {
+			toast.error("Failed to vote! Please try again later", {
+				duration: CONFIG.TOAST_ERROR_DURATION_MS,
+			})
+		},
+	})
+
+	if (isLoading) {
+		return (
+			<div className="relative">
+				<LoadingPage />
+			</div>
+		)
+	}
+
+	const handleNavigateToPost = (postId: string, authorUsername: string) => {
+		// preventing navigate when user selecting text e.g post content text
+		if (!window.getSelection()?.toString()) {
+			router
+				.push(`/post/${authorUsername}/status/${postId}`)
+				.catch(() => toast.error("Error while navigation to post"))
+		}
+	}
+
+	const handlePostClick = (clickCapture: ClickCapture, postsWithUser: PostWithAuthor) => {
+		const { post } = postsWithUser
+
+		switch (clickCapture.action) {
+			case "navigation":
+				handleNavigateToPost(post.id, postsWithUser.author.username)
+				break
+			case "deletePost":
+				deletePost.mutate(post.id)
+				break
+			default:
+				toast.error("Error while post click", {
+					duration: CONFIG.TOAST_ERROR_DURATION_MS,
+				})
+				break
+		}
+	}
+
+	const openDialog = quotePopUp != null && user != null
 
 	return (
 		<Layout
@@ -141,42 +199,88 @@ const Home: NextPage<{ user: User; usersToFollow: UserToFollow[] }> = ({ user, u
 			}
 		>
 			<div className="pt-2">
-				<div className="flex">
-					<ProfileAvatarImageUrl src={user.profileImageUrl} />
-					<div className="w-full pl-1">
-						<input
-							className="w-full rounded-xl border-2 border-solid p-1 text-lg outline-none"
-							placeholder={
-								postContent?.poll ? "Ask a question!" : "What is happening?!"
-							}
-							onChange={(e) =>
-								setPostContent({ ...postContent, message: e.target.value })
-							}
-							type="text"
-						></input>
-						{postContent.poll && (
-							<CreatePoll
-								onRemovePoll={handleRemovePoll}
-								pollChoicesDispatch={pollChoicesDispatch}
-								pollLengthDispatch={pollLengthDispatch}
-								pollLength={pollLengthState}
-								choices={pollChoicesState}
+				<ConnectorCreatePost
+					profileImageUrl={user.profileImageUrl}
+					refetch={() =>
+						void (async () => {
+							await refetch()
+						})
+					}
+				/>
+				<h1 className="p-2 text-2xl font-semibold">Your last posts:</h1>
+				<div>
+					<ul ref={ulRef}>
+						{posts?.map(({ author, post, signInUser }) => (
+							<PostItem
+								key={post.id}
+								postWithUser={{ author, post, signInUser }}
+								onClickCapture={(clickCapture) => {
+									handlePostClick(clickCapture, { author, post, signInUser })
+								}}
+								menuItemsType={getPostMenuItemsType(
+									signInUser?.authorFollowed ?? false,
+									user,
+									author.id
+								)}
+								footer={
+									<PostFooter
+										isForwarded={signInUser?.isForwarded ?? false}
+										onForwardClick={() => {
+											if (signInUser?.isForwarded ?? false) {
+												removePostForward.mutate(post.id)
+											} else {
+												forwardPost.mutate(post.id)
+											}
+										}}
+										onLikeClick={() => {
+											if (signInUser?.isLiked ?? false) {
+												unlikePost.mutate(post.id)
+											} else {
+												likePost.mutate(post.id)
+											}
+										}}
+										onQuoteClick={() => {
+											setQuotePopUp({ author, post, signInUser })
+										}}
+										sharedCount={post.quotedCount + post.forwardsCount}
+										isLiked={signInUser?.isLiked ?? false}
+										likeCount={post.likeCount}
+										username={author.username}
+										replyCount={post.replyCount}
+										postId={post.id}
+									/>
+								}
+							>
+								<PostContentSelector
+									postWithAuthor={{ author, post, signInUser }}
+									pollVote={(choiceId) =>
+										pollVote.mutate({
+											postId: post.id,
+											choiceId,
+										})
+									}
+								/>
+							</PostItem>
+						))}
+					</ul>
+					<dialog open={openDialog}>
+						{quotePopUp && user && (
+							<PostQuotePopUp
+								author={quotePopUp.author}
+								createdAt={quotePopUp.post.createdAt}
+								message={quotePopUp.post.content}
+								onCloseModal={() => setQuotePopUp(null)}
+								onPostQuote={() => {
+									quotePost.mutate({
+										content: quoteMessage ?? "",
+										quotedPostId: quotePopUp.post.id,
+									})
+								}}
+								onMessageChange={(message) => setQuoteMessage(message)}
 							/>
 						)}
-					</div>
+					</dialog>
 				</div>
-				<footer className="ml-16 flex">
-					<div className="flex p-2" onClick={handlePollIconClick}>
-						<Icon iconKind="poll" />
-					</div>
-					<div className="w-full"></div>
-					<div className="mr-2">
-						<PrimaryButton onClick={handleCreatePost}>Post</PrimaryButton>
-					</div>
-				</footer>
-
-				<h1 className="p-2 text-2xl font-semibold">Your last posts:</h1>
-				<HomeFetchPosts isUserFollowProfile={null} user={user} userId={user.id} />
 			</div>
 		</Layout>
 	)
